backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
pydantic[email]
python-multipart
